Discard compass meta with a null world
Discard compass meta with a null world as setting a null world in compass meta is not possible on regular Bukkit servers. A null world indicates that the world for which the location was set does not exist, so the coordinates won't have a useful meaning in any currently existing world.