Add fixed_array, array_fill.
A fixed_array is an array type which maps very closely to the underlying
system. This is not meant for programming high level concepts, in which
case the existing array type is sufficient and preferred. Instead, this
is a strongly typed (for now not using generics) fixed length array.
The data type doesn't extend CArray, nor is it intended to be integrated
with most of the API. It's meant only for mapping to low level concepts.
As the need arises, it may be prudent to create mappings from the normal
array type to fixed_array and vice versa, but only in special cases will
it be integrated into the API directly. More generically, it does extend
Iterable however, and so any code which works directly with arrays but
not Iterable/ArrayAccess should be converted to use these interfaces
instead of integrating with CFixedArray.

This will however be instrumental in creating performant system code
while implementing the native compiler, and will also be useful in the
future in regards to creating integrations with for instance C code.

array_fill is also provided, which does integrate with fixed_array.