Implment DocumentSymbols in langserv (outline view).
This allows for links to procs, binds, and aliases to be added to the
outline view in the IDE. Special support is added for aliases, but for
pure ms files, it's simply a matter of the function extending
DocumentSymbolProvider and returning the label for the item as well as
the kind (which determines the icon). No changes to the VSC extension
are required for this change to work.