Fix LangServ compile errors/warnings being added to the wrong LangServ file
Fix LangServ compile errors and warnings being displayed in the file that has caused the compilation of the file containing the actual errors and warnings. This simply filters those errors and warnings, which is a good temporary solution until LangServ supports some whole-project setup that works well with multi-file static analysis.