Add first version of deep bind compilation checks. (#1301)
* Add Declarative Prefilter pattern

This changes the format of the prefilters from a procedural
to a declarative format. This makes specifying the event a bit more
cumbersome, but at a huge payoff. For one, the docs and filtering are
harder to screw up, but primarily, this should be a substantially better
performing pattern. Previously, many of the events used a
PrefilterNonMatchException to signal that the prefilter doesn't match.
This was less code, but throwing exceptions is substantially more
expensive than simply returning a false boolean. However, doing a
procedural approach and only dealing with booleans would be even more
code than this new approach is, so this is a good balance between
performance and development effort.

Additionally, this allows hooking into the compilation process at a deeper
level, meaning that hardcoded prefilters can be typechecked and deeply
inspected for correctness, and in combination with the LangServ changes,
makes for much better error detection. For now, this part of the feature is
only enabled when static analysis is specifically enabled, since otherwise
the typecheck method isn't called, and thus the error checking doesn't
occur. A future PR can enable this by default (even if general errors remain
suppressed) but that has some prerequisites before that can be done.

Remaining work to be done is to obviously convert every event, but this
can be done at a slower pace in the future. Documentation in the docs
object is overwritten by the docs in the Prefilters, but for now, the
brackets should remain as they are, though they can be emptied out.

Co-authored-by: Pieter12345 <P.J.S.Kools@student.tudelft.nl>