Make clear_task(id) also clear timeouts in TaskManager
Previously when using clear_task(id) with an id from a timeout, it wouldn't execute but it would still be retained by the TaskManager. So this commit changes it so that clear_task(id) uses the TaskManager to clear timeouts.

This also changes the killTaskRunnable for timeout tasks to reflect how they would be cleared: prior to executing, not during execution.