Resolve typename before returning up further.
This is required for functions that return a typename, because the
return type of `T get_value<T>(...)` is not actually `T`, it's whatever
type the user specified (or implied) in the call. So
`get_value<int>(...)`'s return type is int, which is what the
typechecking needs.

This is currently supported through signatures, but was not being
resolved properly before returning up the stack.