Allow "teeing up" of aggressive deprecations.
The annotation now allows for a deprecation version field, which, if the
current version is less than that, will cause it to do nothing. Then,
when the version is bumped up, it will automatically start to be
aggressively deprecated. This allows for deprecations to be made at any
point in the release cycle, without actually changing the behavior, but
then on version bump, all the deprecations automatically take effect.