Make constants their actual type in strict mode, auto otherwise.
This changes the declared type (not to be confused with the actual type)
of constants to auto in non-strict mode, and their actual types in
strict mode. This is necessary, because currently, code such as `'123' >
10` is supported, and needs to continue to be supported into the future.
However, adding typechecking to functions will start to break this code,
which is undesireable. Therefore, in non-strict mode, constants will be
declared as type auto.

Additionally, for functions that end up matching multiple signatures,
instead of returning the type union for possible return values, it now
returns auto, if any of the parameters are auto. This prevents further
type errors up the chain when signatures might be ambiguous. For
instance, if we have both array func(int) and int func(string), then in
non-strict mode, `func(@auto)` would be ambiguous. However, since the
parameter is auto, the user has already indicated that they are ok with
deferring typechecking until runtime anyways, so we prevent requiring a
cast in this case by returning auto, so `array @a = func(@auto)` will
also continue to work with no casts. This behavior is the same in both
strict and non-strict mode, but is more likely to be triggered in
non-strict mode due to the fact that constants are auto.

A few bugs were fixed with strict mode and centry, as well as when
unknonwn functions are found in certain cases.