Full support for statements.
switch(_ic) has been rewritten to use the early keyword processor,
which makes the implementation easier and more robust. This was broken
when adding full support for statements. Additionally, the late keyword
processor was changed to allow no argument keywords, such as with
`return;`, which was not working when followed by dead code.