Don't nest statements, and add back statement checks.
As a modification to how it was before, something like `if(@v;){}` will
now pull up the value to be `if(@v){}` in non strict mode, and in strict
mode, will throw an error. This is necessary, because otherwise this is
a behavior change, since before those were equivalent statements. We
cannot warn/error in non-strict mode, because we still have to support
actual sconcat, and the logic to determine that is too difficult for
now, but larger changes could allow that in the future. In any case, in
such a circumstance, we now simply pull the original value up out of the
statements call, which restores original behavior.

Additionally, statements is supposed to have multiple values. Those are
now separate values in the top level statements call, rather than
nested statements.