Add execution of anything with parenthesis.
Previously, ivars were special cased when executing them such as `@a()`.
This caused the compiler to treat `@a` as a FUNC_NAME token, and then
the runtime had special handling for if the CFunction started with `@`.
Now, however, this is handled at compile time, by rewriting statements
that end with parenthesis into execute. Whatever the previous node was
will now as a whole work to execute, not just for ivars, including
chained executions such as `@a()()`.

There is currently one exception, which is hardcoded to stop being an
exception in 3.3.7, which is that if the left parenthesis is on a
different line than the token being executed, the rewrite is not done,
and instead a warning is issued. It can be fixed by moving the
parenthesis up (if execution is desired) or use of a semicolon (if
execution is not desired).