Add automatic tab completion to @commands.
This provides reasonable defaults for commands depending on the name of
the variable and any constants, but in addition to proper types provided
in @param elements, will provide player name, boolean, adhoc enum, and proper
enum type completions automatically.

Also introduces a CNativeClosure, which can return a Callable which runs
native java code.