Add SelfStatements.
This forces some functions to always be treated as statements, such as
while, which functions as its own statement. This has two effects, one,
a semi-colon is not needed, and in fact using one may cause an empty
statement suggestion to appear. Secondly, when used as parameters to
other functions, this can highlight errors with code, since by
definition, they all return void, and should not be used as inputs.
(Currently an exception is made for sconcat, and non-strict mode in
general.)

In the future, when strict mode checks for missing statements, this will
bring the syntax in line with expectations for modern languages.
Further, this simplifies the rewritten AST, as more things end up being
pulled up into __statements__, rather than being an additional
non-statement node, which can cause an sconcat to appear when
unexpected. Additionally, as these functions are proper statements,
this also prevents accidental execution when using parenthesis, leading
to perhaps less runtime errors being introduced after the parenthesis execution
change.

If, switch, and switch_ic are currently the only "complex" functions,
where they may or may not be statements, depending on their argument
set. The rules for each of these is that if the "return" values are
statements, then they are also considered statements, and if none of the
"return" values are statements, then they are not considered statements.
This allows for them to be used as arguments where expected, but will
help catch unexpected situations at compile time (if strict mode is on).