Improve script eval performance
This reduces Script eval time as well as any other places that use getEnv(). In Script eval this reduces environment hashmap lookups by at least ten per function call, though I'm guessing the average is closer to twelve depending on the function and how many ivars are passed to it. This also optimizes out the creation of an ArrayList and another array. Including a couple smaller optimizations, this ultimately reduced total operation time by 33% in a real script used to test the results.