Fix email by updating its dependency
Email sending did not work in our setup anymore.
And apparently updating the dependency fixes that.

Now the artifact/group is a different one.
With the difference that 'javax.mail' has not been updated since 2013, while 'com.sun.mail' was updated last in 2018 and has a higher version number.

Both seems to stem from the same original source. As the package in both are 'com.sun.mail'.

I removed the maven shade relocation for 'javax.mail'. It never did do anything, since package 'javax.mail' never existed, the dependency used 'com.sun.mail' as package.
I assume it was an oversight, that this relocation was added in the first place.

There is an old relocation for 'com.sun.mail' with a comment, that it cannot be relocated due to functionality.
I left that there, as that probably still applies.

The error we got was: "Could not connect to SMTP host: <host>, port <port>"