Exclude some sqlite native libraries from the jar
These operating systems or architectures are not officially supported, are probably unused, are only relevent to cmdline, and most were only recently added to the sqlite package. But they still contribute significantly to the overall jar size. These exclusions reduces file size by around 7.8 MB. A fallback was added so that a user can still use sqlite on these platforms.