Initialize extensions earlier during plugin load
Moves extension initialization to the plugin's onLoad() method. This enables extensions to process things earlier in the loading process using the new extension onLoad() method.

Moves version assignment earlier so that it can be used for upgrade tasks. This is much faster than loadSelfVersion(), saving tens of milliseconds on the load time.