Swap from using openapi directly to using TypeSpec.
This changes the format of the apps.methodscript.com specification to
use TypeSpec. It is an easier to use format, which has several long term
advantages over using OpenAPI yaml directly. There are no changes to the
client itself, other than to provide an actual object for
BuildArtifacts, which simplifies the parsing of the Updater code. The
SwaggerGenerator version is also updated to the latest version of 3.0.0
now.