Add x_get_lock.
This function works in coordination with sychronized(), to provide a
more flexible solution when coordinating between main thread tasks and
custom thread tasks. When using sychronized() on the main thread, this
is very dangerous, and can very easily lead to deadlocks when using
x_run_on_main_thread_now within the secondary thread, which is a common
pattern. x_get_lock provides an automated mechanism to help balance
these concerns, and automates return of control to the main thread, and
a backoff, so that deadlocks become far easier to avoid.

This works in conjunction with the same lock pool as synchronized(), but
it is changed to use a ReentrantLock object, rather than the native java
synchronized() operator.