Remove optional minus sign in hex/bin/oct/dec number regexes
Minus signs in front of these literals in code are handled elsewhere. Cases where this optional minus sign are present will always return in a `NumberFormatException` due to the substring calls not accounting for it. This only seems to affect `xml_ready()`, which calls the method directly.
Fixes errors in core by for example: `xml_read('<a>-0xFF</a>', 'a')`.