Fix core error creating item with non-item material
Until recently, creating an item stack with a material that is not an item type would work but act like an empty item stack when added to an inventory. Paper now validates if it's an item type on creation. This makes CH throw an exception on invalid item types, but continues to convert legacy block-only items to air. material_info() can now be used to check if a material "isItem".