Fix possible thread leak in PN when recompiled
Bug affected the Persistence Network when it was configured to use a file data source (yml, json, ini, csv, xml). If a recompile is followed by garbage collection, a thread leak may occur. Setting corePoolSize to zero allows the thread to timeout, and ultimately the whole pool to be garbage collected.