Add matrix functions.
This adds support for basic matrix math. In general, doing this natively
should be much faster, as the underlying data remains a native java
array of doubles, so no wrapping and unwrapping needs to occur for the
various math operations. This library may be expanded in the future to
cover more linear algebra scenarios, but the basic functionality is in
place now.